+++ fusion/src/Fusion/python/helpers.h
@@ -12,7 +12,7 @@ namespace py = pybind11;
 
 namespace tensor_py_helpers {
 
-inline py::array_t<double> tensor_to_numpy(const Tensor<double> &t) {
+inline py::array_t<float> tensor_to_numpy(const Tensor<float> &t) {
   // Grab the shape vector
   const auto &shape = t.shape_;
   size_t ndim = shape.size();
@@ -21,19 +21,19 @@ inline py::array_t<double> tensor_to_numpy(const Tensor<double> &t) {
   // Build Python-side shape and stride arrays
   std::vector<ssize_t> py_shape(shape.begin(), shape.end());
   std::vector<ssize_t> py_strides(ndim);
-  // C‐contiguous: stride of last dim is sizeof(double)
-  ssize_t running = sizeof(double);
+  // C‐contiguous: stride of last dim is sizeof(float)
+  ssize_t running = sizeof(float);
   for (int i = ndim - 1; i >= 0; --i) {
     py_strides[i] = running;
     running *= static_cast<ssize_t>(shape[i]);
   }
 
   // Allocate the array
-  py::array_t<double> arr(py_shape, py_strides);
+  py::array_t<float> arr(py_shape, py_strides);
   auto buf = arr.request();
-  double *dst = static_cast<double *>(buf.ptr);
+  float *dst = static_cast<float *>(buf.ptr);
 
-  // Copy from our flat std::vector<double>
+  // Copy from our flat std::vector<float>
   const auto &src = t.raw_data();
   if (src.size() != total) {
     throw std::runtime_error("tensor_to_numpy: size mismatch");
